backend/chatbot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
google-generativeai==0.3.2
sentence-transformers==2.2.2
torch==2.1.0
python-multipart==0.0.6
pydantic==2.5.0
openpyxl==3.1.2
pyarrow==14.0.1
requests==2.31.0
python-dotenv==1.0.0
fastapi-cors==0.0.6
pymongo==4.4.0
faiss-cpu==1.7.4

# Development dependencies
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.26.0

# Optional: For advanced features
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1